# Optimized queries in api_util