This is an update of #498.

I've gone through all the queries except for `get_search_term_examples` (as this one wasn't bugged, but I'll revisit it).

I looked at what the code did and noticed that in reality a LOT of data was being passed over the wire only to then be thrown away...

In my testing, on my dataset with ~10,000 photos:

<img width="218" alt="image" src="https://user-images.githubusercontent.com/864376/169172813-9c74c038-6cc7-452a-b45e-6e69b06554fd.png">

| Endpoint | Before | After  |
|----------------------|--------|--------|
| `api/locationtimeline` | 6.34s  | 265 ms |
| `api/locclust` | 2.85s  | 383 ms |
| `api/locationsunburst` | 6.78s  | 424ms  |

Non-scientific of course, I refreshed a couple of times so that stuff was nicely cached and the wrote down the measurement in the Developer tools.

Now, caveats:

The data is 95% identical. But that doesn't mean that the new version is incorrect. It's just that ordering differs between Python and Postgres, and that if you have 2 locations `United States`, the order isn't defined between the 2. So between runs of the application the order might differ. 

Anyway, let's look at the differences:

### `api/locationtimeline` 
I removed the `color` (it's auto generated) and then there is ZERO difference between before and after. 

### `api/locclust`
2 overall differences here:
1) There seems to be a case where location `FOOBAR` has multiple `center` points, and in Postgres a different row gets selected than in Python. In the grand scheme of things, that's ok. 
E.g.:

<table>
<tr>
<th>Old</th><th>New</th>
</tr>
<tr>
<td>

```json
[
    34.212642,
    -118.765539,
    "Albertson Motorway"
]
```
</td>
<td>

```json
[
    34.2131066937852,
    -118.77435513389975,
    "Albertson Motorway"
]
```
</td>
</tr>
</table>

Looking at the data this is based on

```
❯ docker exec -it postgres psql -U librephotos
psql (14.3 (Debian 14.3-1.pgdg110+1))
Type "help" for help.

librephotos=>SELECT
                jsonb_extract_path_text("feature", 'text') "location"
                , jsonb_extract_path_text("feature", 'center', '0')
                , jsonb_extract_path_text("feature", 'center', '1')
            FROM
                "api_photo"
                , jsonb_array_elements(jsonb_extract_path("api_photo"."geolocation_json", 'features')) "feature"
            WHERE (
                "api_photo"."owner_id" = 1
                AND NOT (jsonb_extract_path_text("feature", 'text') ~ '^(-)?[0-9]+$') 
                AND jsonb_extract_path_text("feature", 'text') = 'Albertson Motorway'
            )
            ORDER BY
                "location";
      location      | jsonb_extract_path_text | jsonb_extract_path_text
--------------------+-------------------------+-------------------------
 Albertson Motorway | -118.77435513389975     | 34.2131066937852
 Albertson Motorway | -118.765539             | 34.212642
 Albertson Motorway | -118.76902              | 34.21227
 Albertson Motorway | -118.77778947224327     | 34.21313046635135
 Albertson Motorway | -118.7706               | 34.212324
 Albertson Motorway | -118.76902              | 34.21227
 Albertson Motorway | -118.76946946161694     | 34.21227492691228
 Albertson Motorway | -118.76528957590156     | 34.212641040676544
 Albertson Motorway | -118.76917433989823     | 34.212272051785945
 Albertson Motorway | -118.765539             | 34.212642
 Albertson Motorway | -118.76945875265037     | 34.21227488571074
 Albertson Motorway | -118.77426691188886     | 34.21311322874897
 Albertson Motorway | -118.76529759658607     | 34.212641071525326
 Albertson Motorway | -118.767988             | 34.212391
 Albertson Motorway | -118.767988             | 34.212391
 Albertson Motorway | -118.7679131168037      | 34.21241081036939
 Albertson Motorway | -118.77426691188886     | 34.21311322874897
 Albertson Motorway | -118.767988             | 34.212391
 Albertson Motorway | -118.770701             | 34.212369
 Albertson Motorway | -118.765539             | 34.212642
 Albertson Motorway | -118.785065             | 34.211886
 Albertson Motorway | -118.756957             | 34.211929
 Albertson Motorway | -118.756957             | 34.211929
 Albertson Motorway | -118.756957             | 34.211929
 Albertson Motorway | -118.756957             | 34.211929
 Albertson Motorway | -118.78574210856424     | 34.211788163770834
 Albertson Motorway | -118.7699823            | 34.2122769
 Albertson Motorway | -118.7949393            | 34.2110446
 Albertson Motorway | -118.76835              | 34.212322
 Albertson Motorway | -118.79325245109773     | 34.21076844046635
 Albertson Motorway | -118.76977146673944     | 34.212276088842856
 Albertson Motorway | -118.76990775111406     | 34.21227661318159
 Albertson Motorway | -118.765279             | 34.212641
 Albertson Motorway | -118.76528957590156     | 34.212641040676544
 Albertson Motorway | -118.765279             | 34.212641
(35 rows)

```

We can see that there are differences that resolve to the same main 'point' and that which one gets picked is different in Postgres and in Python.

But this is impossible to reproduce 100% in SQL unless we find out the internal sort order of unsorted things in Python.

2) order of the objects. It doesn't impact the rendered result however. For an explanation as to how the order is different, see the `api/locationsunburst` ordering explanation. 

### `api/locationsunburst`
After zeroing out the colors it is 99% identical.

The difference is in the sort order.

Consider the following letters: A, O, Ö, and Z. In Python sort it would be A, O, Z, Ö.  However in the Postgres sort it is be A, O, Ö, Z because that is where you expect them to be in a natural sort order. 

This exhibits in for example the following cities being sorted differently: 
```diff
{
    "name": "Veszprém",
    "children": [
        {
            "name": "Alsóörs",
            "value": 13,
            "hex": "#000000"
        },
        {
            "name": "Balatonalmádi",
            "value": 3,
            "hex": "#000000"
        },
        {
            "name": "Balatonfüred",
            "value": 32,
            "hex": "#000000"
        },
        {
            "name": "Felsőörs",
            "value": 4,
            "hex": "#000000"
        },
        {
            "name": "Monoszló",
            "value": 17,
            "hex": "#000000"
        },
+       {
+           "name": "Örvényes",
+           "value": 4,
+           "hex": "#000000"
+       },
        {
            "name": "Tapolca",
            "value": 16,
            "hex": "#000000"
        },
        {
            "name": "Tihany",
            "value": 32,
            "hex": "#000000"
        },
        {
            "name": "Veszprém",
            "value": 37,
            "hex": "#000000"
        },
-       {
-           "name": "Örvényes",
-           "value": 4,
-           "hex": "#000000"
-       }
    ],
    "hex": "#000000"
}
```

Let me know what you think!!!