@derneuere 2022-05-19
This is all code from the original codebase, so I can't tell you if there was any intent behind the sorting / data duplication. I would guess that this was probably just a naive implementation.

If you want to debug a specific query/request, you can also just view it in silk. If you are in development mode, then you can access it under /api/silk

Don't worry about the linting, that's on me :)

Awesome work! Also worked on my machine. Merged :)